% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\title{k-means clustering}
\usage{
kmeans_cluster(data, centers, max_iter = 100)
}
\arguments{
\item{data}{- The data object (data frame or matrix) that k-means clustering will be applied to.}

\item{centers}{- The matrix containing the initial centers as computed by kmeans_init.}
}
\value{
- The data object (as a data frame) with a column containing cluster assignments.
}
\description{
k-means clustering
}
\examples{
x <- runif(100, min = 0, max = 10) + rep(c(0, 10), 50)
y <- rnorm(100, 5, 1) + rep(c(0, 10), 50)
data = data.frame(x = x, y = y)

centers <- kmeans_init(data = data, K = 2)
kmeans_cluster(data=data, centers=centers)

}
