% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_report.R
\name{kmeans_report}
\alias{kmeans_report}
\title{Plot data according to cluster assigments}
\usage{
kmeans_report(clustered_data)
}
\arguments{
\item{clustered_data}{-- the data object (data frame or matrix) that k-means clustering has been applied to}
}
\value{
plot -- a scatter plot of the data coloured by cluster assignment (if 2D data)
             -- a data frame reporting points and their cluster assignemnts
}
\description{
Plot data according to cluster assigments
}
\examples{
x <- runif(100, min = 0, max = 10) + rep(c(0, 10), 50)
y <- rnorm(100, 5, 1) + rep(c(0, 10), 50)
data = data.frame(x = x, y = y)

centers <- kmeans_init(data = data, K = 2)
clustered_data <- kmeans_cluster(data=data, centers=centers)
kmeans_report(clustered_data)

}
