% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_report.R
\name{kmeans_report}
\alias{kmeans_report}
\title{Plot data according to cluster assigments}
\usage{
kmeans_report(data, clust_assigns)
}
\arguments{
\item{data}{-- the data object (data frame or matrix) that k-means clustering has been applied to}

\item{clust_assigns}{-- cluster assignments that correspond to each data point}
}
\value{
plot -- a scatter plot of the data coloured by cluster assignment (if 2D data)
             -- a data frame reporting points and their cluster assignemnts
}
\description{
Plot data according to cluster assigments
}
